requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
razorpay
imagekitio
python-multipart
pydantic